Add a new policy to authorize the author user for update and delete recipes in the RecipeController

+++ app/Http/Controllers/Api/RecipeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Models\Recipe;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\RecipeResource;
 use App\Http\Requests\StoreRecipeRequest;
 use App\Http\Requests\UpdateRecipeRequest;
@@ -35,6 +36,8 @@ class RecipeController extends Controller
 
     // update recipe
     public function update(UpdateRecipeRequest $request, Recipe $recipe){
+        Gate::authorize('update', $recipe);
+
         $recipe->update($request->all());
 
         if ($tags = json_decode($request->tags)) {
@@ -46,6 +49,8 @@ class RecipeController extends Controller
 
     // delete recipe
     public function destroy(Recipe $recipe){
+        Gate::authorize('delete', $recipe);
+
         $recipe->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT); // HTTP_NO_CONTENT = 204
